dom-swap-implementation
The current API only permits to render webpages server side, which is cool but that means interactive features cannot be shippied with the html bundle since the server cannot simulate events such as click, mounting ecc.

This branch aims at implementing "svelte3dom", which aims at rendering the js and css bundles meant to be served to the client.

These bundles will swap the the SSR page when it loads, and allows the normal functionality of the svelte page.

This means that your web pages will get the benefit of SSR and allow SEO while at the same time allowing the web page to function properly.

---

NOTE: this branch also fixes the problem of importing ".svelte" files.
Currently ".svelte" files cannot be imported using the "import" keyword, that is because the JS engine does not support custom extensions for the require function (because it's actually officialy deprecated).

Currently on the main branch you can import ".svelte" files using the component("./www/ComponentName.svelte");